% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btn-package.R
\docType{package}
\name{BTN-package}
\alias{BTN}
\alias{BTN-package}
\title{BTN: Functions for handling acoutic telemetry data from the Bergen Telemetry Network}
\description{
Helper functions for managing data for the Bergen Telemetry Network
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertlennox/BTN}
  \item Report bugs at \url{https://github.com/robertlennox/BTN/issues}
}

}
\keyword{internal}
