% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{creator}
\alias{creator}
\title{functions to get receiver and tagging metadata into session}
\arguments{
\item{meta}{is an object brought in from the tagging metadata}

\item{receivers}{is an object brought in from the receivers sheet}
}
\description{
functions to get receiver and tagging metadata into session
}
