% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb.R
\name{read_tb}
\alias{read_tb}
\title{function to read in and manipulate Thelma DB}
\usage{
read_tb(tbdb_file)
}
\arguments{
\item{tbdb_file}{is the name of the thelma database}
}
\description{
function to read in and manipulate Thelma DB
}
